% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryNet.R
\name{as.BinaryNet}
\alias{as.BinaryNet}
\title{Convert a network to either an UndirectedNet or DirectedNet object}
\usage{
as.BinaryNet(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{unused}
}
\value{
a BinaryNet object
}
\description{
Convert a network to either an UndirectedNet or DirectedNet object
}
