% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryNet.R
\name{as.network.Rcpp_DirectedNet}
\alias{as.network.Rcpp_DirectedNet}
\title{Convert an Rcpp_DirectedNet to a network object}
\usage{
\method{as.network}{Rcpp_DirectedNet}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{unused}
}
\value{
a directed network object
}
\description{
Converts a native ernm network into a network object
from the network package.
}
\examples{
edge_list <- matrix(c(1,3),ncol=2)
# create a network with an edge from 1 -> 3
ernm_net <- new(DirectedNet,edge_list,5)

# convert to a network object from the network package
network_net <- as.network(ernm_net)
network_net
}
