% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryNet.R
\name{as.network.Rcpp_UndirectedNet}
\alias{as.network.Rcpp_UndirectedNet}
\title{Convert an Rcpp_UndirectedNet to a network object}
\usage{
\method{as.network}{Rcpp_UndirectedNet}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{unused}
}
\value{
a undirected network object
}
\description{
Convert an Rcpp_UndirectedNet to a network object
}
