% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryNet.R
\name{as.network.UndirectedNet}
\alias{as.network.UndirectedNet}
\title{convert and UndirectedNet to a network object}
\usage{
\method{as.network}{UndirectedNet}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{unused}
}
\description{
convert and UndirectedNet to a network object
}
