% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{methods}
\name{call-symbols}
\alias{call-symbols}
\alias{_rcpp_module_boot_ernm}
\alias{_ernm_initToggles}
\alias{_ernm_initStats}
\alias{initLatent}
\title{Internal symbols}
\description{
Internal symbols used to access compiles code.
}
\keyword{internal}
