% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-methods.R
\name{coef.ernm}
\alias{coef.ernm}
\title{Access ERNM parameters}
\usage{
\method{coef}{ernm}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{unused}
}
\value{
parameter vector
}
\description{
Access ERNM parameters
}
