% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm.R
\name{createCppModel}
\alias{createCppModel}
\title{Creates a model}
\usage{
createCppModel(
  formula,
  ignoreMnar = TRUE,
  cloneNet = TRUE,
  theta = NULL,
  modelArgs = list(modelClass = "Model")
)
}
\arguments{
\item{formula}{the model formula}

\item{ignoreMnar}{ignore missing not at random offsets}

\item{cloneNet}{should the network be cloned}

\item{theta}{the model parameters.}

\item{modelArgs}{additiional arguments for the model, e.g. tapering parameters}
}
\value{
a Model object
}
\description{
Creates a model
}
