% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{class}
\name{DirectedMetropolisHastings}
\alias{DirectedMetropolisHastings}
\alias{Rcpp_DirectedMetropolisHastings}
\title{DirectedMetropolisHastings Class}
\format{
An object of class \code{C++Class} of length 1.
}
\usage{
DirectedMetropolisHastings
}
\description{
A class implementing the Metropolis-Hastings algorithm for directed networks.
}
\section{Methods}{

\describe{
  \item{\code{setModel(model)}}{Sets the model to be used.}
  \item{\code{getModel()}}{Returns the current model.}
  \item{\code{setDyadToggleType(type)}}{Sets the dyad toggle type.}
  \item{\code{generateSample()}}{Generates a network sample.}
  \item{\code{generateSampleStatistics()}}{Computes sample statistics.}
}
}

\examples{
\dontrun{
  dmh <- new(DirectedMetropolisHastings)
  dmh$setModel(myDirectedModel)
  sample <- dmh$generateSample()
}

}
\keyword{datasets}
