% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{data}
\name{DirectedTaperedModel}
\alias{DirectedTaperedModel}
\title{DirectedTaperedModel Class}
\format{
An object of class \code{C++Class} of length 1.
}
\usage{
DirectedTaperedModel
}
\description{
A class representing a tapered model for directed networks.
}
\section{Methods}{

\describe{
  \item{\code{setTau(tau)}}{Sets the tapering parameter tau.}
  \item{\code{tau()}}{Returns the current tau parameters.}
  \item{\code{setCenters(centers)}}{Sets the center parameters for tapering.}
  \item{\code{centers()}}{Retrieves the center parameters.}
}
}

\keyword{datasets}
