% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.matchParams}
\alias{.matchParams}
\title{Parses evaulated parameters as if in a function call.
using names and positional matching}
\usage{
.matchParams(lis, params)
}
\arguments{
\item{lis}{a list of parameter values}

\item{params}{a named list of default parameters}
}
\description{
Parses evaulated parameters as if in a function call.
using names and positional matching
}
