% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dutch_school}
\alias{dutch_school}
\title{Dutch School Data}
\usage{
data(dutch_school)
}
\description{
Dutch School Data
}
\section{Data taken from https}{
//www.stats.ox.ac.uk/~snijders/siena/tutorial2010_data.htm.
Processed as undirected networks.
}

\references{
Snijders, T.A.B., Steglich, C.E.G., and van de Bunt, G.G. (2010),
Introduction to actor-based models for network dynamics, Social Networks 32, 44-60,
http://dx.doi.org/10.1016/j.socnet.2009.02.004.
}
