% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{class}
\name{ErnmModels}
\alias{ErnmModels}
\alias{DirectedModel}
\alias{UndirectedModel}
\alias{Rcpp_DirectedModel-class}
\alias{Rcpp_UndirectedModel-class}
\alias{Rcpp_DirectedTaperedModel-class}
\alias{Rcpp_UndirectedTaperedModel-class}
\alias{DirectedTaperedModel}
\alias{UndirectedTaperedModel}
\title{CPP model objects}
\description{
CPP backend for model objects used internally in ernm functions. `ErnmModels` covers ERNM's native, Rcpp-module model classes exposing their methods to the R interface.
}
