% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ernmPackageSkeleton}
\alias{ernmPackageSkeleton}
\title{Create an ERNM package skeleton}
\usage{
ernmPackageSkeleton(path = ".")
}
\arguments{
\item{path}{A character string specifying the directory where the package skeleton will be created.}
}
\value{
A logical value indicating whether the copy was successful.
}
\description{
Creates a skeleton for a package extending the ernm package by copying an example package.
}
