% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{class}
\name{ErnmSamplers}
\alias{ErnmSamplers}
\alias{Rcpp_DirectedMetropolisHastings-class}
\alias{Rcpp_UndirectedMetropolisHastings-class}
\alias{Rcpp_UndirectedCdSampler-class}
\alias{Rcpp_DirectedCdSampler-class}
\alias{Rcpp_UndirectedGibbsCdSampler-class}
\alias{Rcpp_DirectedGibbsCdSampler-class}
\alias{Rcpp_UndirectedGibbsCdSampler2-class}
\alias{Rcpp_DirectedGibbsCdSampler2-class}
\alias{DirectedMetropolisHastings}
\alias{UndirectedMetropolisHastings}
\alias{UndirectedCdSampler}
\alias{DirectedCdSampler}
\alias{UndirectedGibbsCdSampler}
\alias{DirectedGibbsCdSampler}
\alias{UndirectedGibbsCdSampler2}
\alias{DirectedGibbsCdSampler2}
\title{Metropolis samplers}
\description{
Metropolis hastings CPP samplers, used in the fitting procedure for ernm models.
}
