% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-models.R
\name{FullErnmModel}
\alias{FullErnmModel}
\title{creates an ERNM likelihood model}
\usage{
FullErnmModel(sampler, logLik, ...)
}
\arguments{
\item{sampler}{a sampler}

\item{logLik}{a log likelihood function (optional)}

\item{...}{additional parameters for the log likelihood}
}
\description{
creates an ERNM likelihood model
}
