% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mcmcse}
\alias{mcmcse}
\title{MCMC standard error by batch}
\usage{
mcmcse(x, expon = 0.5)
}
\arguments{
\item{x}{a statistic vector}

\item{expon}{batch size control}
}
\description{
MCMC standard error by batch
}
