% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-models.R
\name{MissingErnmModel}
\alias{MissingErnmModel}
\title{creates an ERNM likelihood model}
\usage{
MissingErnmModel(observedSampler, unobservedSampler, ...)
}
\arguments{
\item{observedSampler}{a sampler}

\item{unobservedSampler}{a sampler conditional upon the observed values}

\item{...}{additional parameters for the log likelihood}
}
\description{
creates an ERNM likelihood model
}
