% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryNet.R
\name{plot.DirectedNet}
\alias{plot.DirectedNet}
\title{plot an DirectedNet object}
\usage{
\method{plot}{DirectedNet}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional parameters for plot.network}
}
\value{
No return value, invisibly NULL
}
\description{
plot an DirectedNet object
}
