% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-fit.R
\name{plot.ernm}
\alias{plot.ernm}
\title{plot an ernm object}
\usage{
\method{plot}{ernm}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{unused}
}
\value{
No return value, plots the likelihood history
}
\description{
plot an ernm object
}
