% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryNet.R
\name{plot.Rcpp_UndirectedNet}
\alias{plot.Rcpp_UndirectedNet}
\title{Plot an UndirectedNet object}
\usage{
\method{plot}{Rcpp_UndirectedNet}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional parameters for plot.network}
}
\value{
No return value, invisibly NULL
}
\description{
Plot an UndirectedNet object
}
\examples{

# create ring network with 5 vertices
edge_list <- matrix(c(1,2,2,3,3,4,4,5,5,1),ncol=2, byrow=TRUE)
ernm_net <- new(UndirectedNet,edge_list,5)

# basic plot
plot(ernm_net)

# change vertex point size (see plot.network)
plot(ernm_net, vertex.cex=.5)

}
