% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryNet.R
\name{plot.UndirectedNet}
\alias{plot.UndirectedNet}
\title{plot an UndirectedNet object}
\usage{
\method{plot}{UndirectedNet}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional parameters for plot.network}
}
\description{
plot an UndirectedNet object
}
