% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-methods.R
\name{print.ernm}
\alias{print.ernm}
\title{Print ernm object}
\usage{
\method{print}{ernm}(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{unused}
}
\value{
No return value, prints summary
}
\description{
Print ernm object
}
