% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-methods.R
\name{print.ErnmSummary}
\alias{print.ErnmSummary}
\title{Print a ERNM summary object}
\usage{
\method{print}{ErnmSummary}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{parameters passed to print.data.frame}
}
\value{
x invisibly
}
\description{
Print a ERNM summary object
}
