% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_registration.R
\name{register_rcpp_net_methods}
\alias{register_rcpp_net_methods}
\title{Register setter methods for Rcpp net objeccts}
\usage{
register_rcpp_net_methods()
}
\value{
no return value
}
\description{
Register setter methods for Rcpp net objeccts
}
