% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{runErnmCppTests}
\alias{runErnmCppTests}
\title{runErnmCppTests}
\value{
A logical value indicating whether all tests passed.
}
\description{
Runs the internal C++ tests for the ernm package.
}
\examples{
\dontrun{
  runErnmCppTests()
}

}
