% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samplike}
\alias{samplike}
\title{Sampson's Monks Data}
\usage{
data(samplike)
}
\description{
Sampson's Monks Data
}
\section{NOTE COPIED FROM ERGM PACKAGE}{

Mislabeling in Versions Prior to 3.6.1: In \code{ergm} version
3.6.0 and earlier, the adjacency matrices of the datasets
reflected an older ordering of the names.
}

\references{
White, H.C., Boorman, S.A. and Breiger, R.L. (1976).
\emph{Social structure from multiple networks. I. Blockmodels of roles and
positions.} American Journal of Sociology, 81(4), 730-780.
}
\seealso{
florentine, network, plot.network, ergm
}
