% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm.R
\name{simulateStatistics}
\alias{simulateStatistics}
\title{Simulate statistics}
\usage{
simulateStatistics(
  formula,
  theta,
  nodeSamplingPercentage = 0.2,
  mcmcBurnIn = 10000,
  mcmcInterval = 100,
  mcmcSampleSize = 100,
  ignoreMnar = TRUE,
  modelArgs = list(modelClass = "Model"),
  ...
)
}
\arguments{
\item{formula}{the model formula}

\item{theta}{model parameters}

\item{nodeSamplingPercentage}{how often the nodes should be toggled}

\item{mcmcBurnIn}{burn in}

\item{mcmcInterval}{interval}

\item{mcmcSampleSize}{sample size}

\item{ignoreMnar}{ignore missing not at random offsets}

\item{modelArgs}{additiional arguments for the model, e.g. tapering parameters}

\item{...}{additional arguments to createCppSampler}
}
\value{
a list of statistics
}
\description{
Simulate statistics
}
