% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-methods.R
\name{summary.ernm}
\alias{summary.ernm}
\title{Summary for ernm object}
\usage{
\method{summary}{ernm}(object, include_AIC = TRUE, ...)
}
\arguments{
\item{object}{object}

\item{include_AIC}{whether to include AIC in the summary, will slow down}

\item{...}{unused}
}
\value{
a data frame summary of the model
}
\description{
Summary for ernm object
}
