% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-models.R
\name{TaperedModel}
\alias{TaperedModel}
\title{creates an ERNM likelihood model}
\usage{
TaperedModel(sampler, ...)
}
\arguments{
\item{sampler}{a sampler}

\item{...}{additional parameters for the log likelihood}
}
\description{
creates an ERNM likelihood model
}
