% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{data}
\name{UndirectedGibbsCdSampler}
\alias{UndirectedGibbsCdSampler}
\title{UndirectedGibbsCdSampler Class}
\format{
An object of class \code{C++Class} of length 1.
}
\usage{
UndirectedGibbsCdSampler
}
\description{
A class implementing a Gibbs-based CdSampler for undirected networks.
}
\section{Methods}{

\describe{
  \item{\code{setModel(model)}}{Sets the model to be sampled.}
  \item{\code{getModel()}}{Returns the current model.}
  \item{\code{setDyadToggleType(type)}}{Sets the dyad toggle type.}
  \item{\code{setVertexToggleType(type)}}{Sets the vertex toggle type.}
  \item{\code{setDyadProbability(prob)}}{Sets the probability for dyad toggling.}
  \item{\code{generateSample()}}{Generates a network sample.}
  \item{\code{generateSampleStatistics()}}{Computes sample statistics.}
}
}

\keyword{datasets}
