% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{data}
\name{UndirectedTaperedModel}
\alias{UndirectedTaperedModel}
\title{UndirectedTaperedModel Class}
\format{
An object of class \code{C++Class} of length 1.
}
\usage{
UndirectedTaperedModel
}
\description{
A class representing a tapered model for undirected networks.
}
\section{Methods}{

\describe{
  \item{\code{setTau(tau)}}{Sets the tapering parameter.}
  \item{\code{tau()}}{Retrieves the current tau parameter(s).}
  \item{\code{setCenters(centers)}}{Sets the center parameters.}
  \item{\code{centers()}}{Retrieves the center parameters.}
}
}

\keyword{datasets}
