% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-fit.R
\name{vcov.ernm}
\alias{vcov.ernm}
\title{parameter covariance matrix}
\usage{
\method{vcov}{ernm}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{unused}
}
\value{
covariance matrix
}
\description{
parameter covariance matrix
}
