% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryNet.R
\name{as.BinaryNet}
\alias{as.BinaryNet}
\title{Convert a network to either an Rcpp_UndirectedNet or Rcpp_DirectedNet object}
\usage{
as.BinaryNet(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{unused}
}
\value{
either an Rcpp_UndirectedNet or Rcpp_DirectedNet object
}
\description{
Convert a network to either an Rcpp_UndirectedNet or Rcpp_DirectedNet object
}
\examples{

data(samplike)

# convert Sampson's monks into a native ernm network
net <- as.BinaryNet(samplike)
net[["group"]]
net[1:5,1:5]

}
