% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm.R
\name{calculateStatistics}
\alias{calculateStatistics}
\title{calculate model statistics from a formula}
\usage{
calculateStatistics(formula)
}
\arguments{
\item{formula}{An ernm formula}
}
\value{
a list of statistics
}
\description{
calculate model statistics from a formula
}
