% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm.R
\name{calculateStatistics}
\alias{calculateStatistics}
\title{Calculate network statistics}
\usage{
calculateStatistics(formula)
}
\arguments{
\item{formula}{An ernm formula (see \code{\link{ernm-formula}})}
}
\value{
a named vector of statistics
}
\description{
Calculates all network statistics
}
\examples{
data(samplike)
calculateStatistics(samplike ~ edges() +
 nodeCount("group") +
 nodeMatch("group") +
 homophily("group") +
 triangles())
}
