% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm.R
\name{createCppSampler}
\alias{createCppSampler}
\title{create a sampler}
\usage{
createCppSampler(
  formula,
  modelArgs = list(modelClass = "Model"),
  dyadToggle = NULL,
  dyadArgs = list(),
  vertexToggle = NULL,
  vertexArgs = list(),
  nodeSamplingPercentage = 0.2,
  ignoreMnar = TRUE,
  theta = NULL,
  ...
)
}
\arguments{
\item{formula}{the model formula}

\item{modelArgs}{additiional arguments for the model, e.g. tapering parameters}

\item{dyadToggle}{the method of sampling to use. Defaults to alternating between nodal-tie-dyad and neighborhood toggling.}

\item{vertexToggle}{the method of vertex attribuate sampling to use.}

\item{nodeSamplingPercentage}{how often the nodes should be toggled}

\item{ignoreMnar}{ignore missing not at random offsets}

\item{theta}{parameter values}
}
\description{
create a sampler
}
