% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data_docs.R
\docType{data}
\name{dutch_school}
\alias{dutch_school}
\title{Dutch School Data}
\format{
An object of class \code{list} of length 4.
}
\source{
Knecht, A. (2004). *Network and actor attributes in early adolescence*. 
DANS Data Station Social Sciences and Humanities. 
DOI: \doi{10.17026/dans-z9b-h2bp}.
Available at: \url{https://doi.org/10.17026/dans-z9b-h2bp}
}
\usage{
data(dutch_school)

data(dutch_school)
}
\description{
This dataset contains network and actor attributes collected in early adolescence.
It is provided by Andrea Knecht and stored in the package.
}
\section{Data taken from https}{
//www.stats.ox.ac.uk/~snijders/siena/tutorial2010_data.htm.
Processed as undirected networks.
}

\references{
Snijders, T.A.B., Steglich, C.E.G., and van de Bunt, G.G. (2010),
Introduction to actor-based models for network dynamics, Social Networks 32, 44-60,
http://dx.doi.org/10.1016/j.socnet.2009.02.004.
}
\keyword{datasets}
