% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-fit.R
\name{ernmFit}
\alias{ernmFit}
\title{fit an ernm model}
\usage{
ernmFit(
  sampler,
  theta0,
  mcmcBurnIn = 10000,
  mcmcInterval = 100,
  mcmcSampleSize = 10000,
  minIter = 3,
  maxIter = 40,
  objectiveTolerance = 0.5,
  gradTolerance = 0.25,
  meanStats,
  verbose = 1,
  method = c("bounded", "newton")
)
}
\arguments{
\item{sampler}{the ErnmModel}

\item{theta0}{initial starting values}

\item{mcmcBurnIn}{burn in}

\item{mcmcInterval}{interval}

\item{mcmcSampleSize}{sample size}

\item{minIter}{minimum number of iterations}

\item{maxIter}{maximum number of iterations}

\item{objectiveTolerance}{convergance criteria on change in log likelihood ratio}

\item{gradTolerance}{convergance criteria on scaled gradient}

\item{meanStats}{if non-missing, these are the target statistics}

\item{verbose}{level of verbosity 0, 1, or 2}

\item{method}{the optimization method to use}
}
\value{
ernm object
}
\description{
fit an ernm model
}
