% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-methods.R
\name{logLik.ernm}
\alias{logLik.ernm}
\title{MCMC approximate log-likelihood}
\usage{
\method{logLik}{ernm}(object, ...)
}
\arguments{
\item{object}{an ernm object}

\item{...}{unused}
}
\description{
MCMC approximate log-likelihood
}
\examples{
\dontrun{
fit <- ernm(samplike ~ edges() + nodeCount("group") + nodeMatch("group") | group)
logLik(fit)
AIC(fit)
BIC(fit)
}

}
