% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mcmcEss}
\alias{mcmcEss}
\title{MCMC Effective Sample Size}
\usage{
mcmcEss(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value representing the effective sample size.
}
\description{
Computes the effective sample size from a statistic vector.
}
