% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mcmcse}
\alias{mcmcse}
\title{MCMC Standard Error by Batch}
\usage{
mcmcse(x, expon = 0.5)
}
\arguments{
\item{x}{A numeric vector of statistics.}

\item{expon}{A numeric value controlling the batch size; default is 0.5.}
}
\value{
A numeric value representing the estimated standard error.
}
\description{
Computes the MCMC standard error from a statistic vector using a batching method.
}
\examples{
x <-

}
