% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-models.R
\name{MissingErnmModel}
\alias{MissingErnmModel}
\title{Creates an ERNM likelihood model with missing data}
\usage{
MissingErnmModel(observedSampler, unobservedSampler, ...)
}
\arguments{
\item{observedSampler}{a C++ sampler}

\item{unobservedSampler}{a C++ sampler conditional upon the observed values}

\item{...}{additional parameters for the log likelihood}
}
\value{
a MarModel object
}
\description{
Creates an ERNM likelihood model with missing data
}
