% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-methods.R
\name{plot.ernm}
\alias{plot.ernm}
\title{Plot an ernm object}
\usage{
\method{plot}{ernm}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{passed to plot function}
}
\value{
No return value, plots the likelihood history
}
\description{
Plot an ernm object
}
