% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryNet.R
\name{plot.Rcpp_DirectedNet}
\alias{plot.Rcpp_DirectedNet}
\title{Plot an Rcpp_UnirectedNet object}
\usage{
\method{plot}{Rcpp_DirectedNet}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional parameters for plot.network}
}
\value{
No return value, invisibly NULL
}
\description{
Plot an Rcpp_UnirectedNet object
}
