% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{data}
\name{runErnmCppTests}
\alias{runErnmCppTests}
\title{runErnmCppTests}
\format{
An object of class \code{C++Function} of length 1.
}
\usage{
runErnmCppTests
}
\value{
A logical value indicating whether all tests passed.
}
\description{
Runs the internal C++ tests for the ernm package.
}
\examples{
\dontrun{
  runErnmCppTests()
}

}
\keyword{datasets}
