% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{runErnmCppTests}
\alias{runErnmCppTests}
\title{runErnmCppTests}
\value{
No return value, invisibly NULL
}
\description{
Runs the internal C++ tests for the ernm package in order. If a particular
test fails it prints an informative error message and stops without running the subsequent tests.
}
\examples{
runErnmCppTests()

}
\keyword{internal}
