% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data_docs.R
\docType{data}
\name{samplike}
\alias{samplike}
\alias{sampson}
\title{Sampson's Monks Data}
\format{
An object of class \code{network} of length 5.
}
\source{
Sampson, S. F. (1969). *Crisis in a cloister*. Unpublished Ph.D. dissertation, Cornell University.
}
\usage{
data(samplike)

data("samplike")
}
\description{
This dataset represents the social network of relationships among monks in a monastery,
as studied by Samuel F. Sampson. The data were collected during a period of instability
and document both positive and negative interactions among the monks.
}
\details{
The study recorded friendships, antagonisms, and other social relationships among the monks
before and after a significant schism occurred in the monastery.
}
\section{NOTE COPIED FROM ERGM PACKAGE}{

Mislabeling in Versions Prior to 3.6.1: In \code{ergm} version
3.6.0 and earlier, the adjacency matrices of the datasets
reflected an older ordering of the names.
}

\references{
White, H.C., Boorman, S.A. and Breiger, R.L. (1976).
\emph{Social structure from multiple networks. I. Blockmodels of roles and
positions.} American Journal of Sociology, 81(4), 730-780.
}
\seealso{
florentine, network, plot.network, ergm
}
\keyword{datasets}
