% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-methods.R
\name{summary.ernm}
\alias{summary.ernm}
\title{Summary for ernm object}
\usage{
\method{summary}{ernm}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{unused}
}
\value{
an ErnmSummary object
}
\description{
Summary for ernm object
}
