% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernm-fit.R
\name{summary.ernm}
\alias{summary.ernm}
\title{summary}
\usage{
\method{summary}{ernm}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{unused}
}
\value{
a data frame summary of the model
}
\description{
summary
}
