% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{taperedErnmLikelihood}
\alias{taperedErnmLikelihood}
\title{(E(g(X)) - g(x_o)^2 for TaperedModel}
\usage{
taperedErnmLikelihood(
  theta,
  centers,
  tau,
  sample,
  theta0,
  stats,
  minEss = 5,
  damping = 0.05
)
}
\arguments{
\item{theta}{parameters}

\item{sample}{mcmc sample}

\item{theta0}{parameter values which generated sample}

\item{stats}{observed statistics}

\item{minEss}{minimum effective sample size}
}
\description{
(E(g(X)) - g(x_o)^2 for TaperedModel
}
